+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Deprecated
 
-- Renamed `parent_child_avatar` to `avatar_pair`.
+- Component `avatar_pair`: renamed `parent_child_avatar` to `avatar_pair`.
+- Components `action_menu` and `select_menu`: renamed `is_right_aligned` to `is_aligned_end` (added RTL support).
 
 ### Updated components